core-shim: add missing int cast on syscall return

+++ core-shim.c
@@ -2791,7 +2791,7 @@ int shim_finit_module(int fd, const char *uargs, int flags)
 
 	return finit_module(fd, uargs, flags);
 #elif defined(__NR_finit_module)
-	return syscall(__NR_finit_module, fd, uargs, flags);
+	return (int)syscall(__NR_finit_module, fd, uargs, flags);
 #else
 	return shim_enosys(0, fd, uargs, flags);
 #endif
@@ -2808,7 +2808,7 @@ int shim_delete_module(const char *name, unsigned int flags)
 
 	return delete_module(name, flags);
 #elif defined(__NR_delete_module)
-	return syscall(__NR_delete_module, name, flags);
+	return (int)syscall(__NR_delete_module, name, flags);
 #else
 	return shim_enosys(0, name, flags);
 #endif
